bbq:test generator
