+++ lib/generators/bbq/install_generator.rb
@@ -1,7 +1,8 @@
+require 'bbq/generator'
+
 module Bbq
   class InstallGenerator < Rails::Generators::Base
-    source_root File.expand_path('../templates', __FILE__)
-    class_option :test_framework, :alias => "-t", :type => :string, :default => "test_unit", :desc => "Test framework to be invoked"
+    include Bbq::Generator
 
     def create_directory
       empty_directory test_root
@@ -14,21 +15,6 @@ module Bbq
 
     def show_readme
       readme "README.#{test_framework_short}" if behavior == :invoke
-    end
-
-    private
-    def test_framework_short
-      case options.test_framework
-      when :test_unit then 'test'
-      end
-    end
-
-    def test_root
-      "#{test_framework_short}/acceptance"
-    end
-
-    def support_root
-      "#{test_root}/support"
     end
   end
 end
