+++ lib/generators/bbq/templates/bbq_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class <%= name.classify %>Test < Bbq::TestCase
+  scenario "the truth" do
+    assert true
+  end
+end
